Filter to Problem File
Added a filter to pull anything that was between 85473 and 107422 into a new file name problem.
This will work for parent features but there is a problem that the children of those parents may not fall within this range so could be left in the file.  
Hopefully the GFF3 viewer would not allow these....
Might be able to work in something to deal with them later, but right now beyond me.